chatbot/version-1/requirements.txt
numpy==1.24.4
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
torchvision==0.16.0+cpu

--extra-index-url https://download.pytorch.org/whl/cpu
pymilvus==2.3.4
sentence-transformers==2.2.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.11.6


# ----- Installs -------

# (venv) bryson@DESKTOP-OFT75GU:~/aiaas-agent$ pip freeze
    # argon2-cffi==23.1.0
    # argon2-cffi-bindings==21.2.0
    # backports-datetime-fromisoformat==2.0.3
    # certifi==2025.4.26
    # cffi==1.17.1
    # charset-normalizer==3.4.2
    # click==8.2.1
    # cryptography==45.0.2
    # environs==9.5.0
    # filelock==3.18.0
    # fsspec==2025.5.1
    # grpcio==1.58.0
    # hf-xet==1.1.2
    # huggingface-hub==0.32.0
    # idna==3.10
    # Jinja2==3.1.6
    # joblib==1.5.1
    # lxml==5.4.0
    # MarkupSafe==3.0.2
    # marshmallow==4.0.0
    # minio==7.2.15
    # mpmath==1.3.0
    # networkx==3.4.2
    # nltk==3.9.1
    # numpy==1.24.4
    # packaging==25.0
    # pandas==2.2.3
    # pdfminer.six==20250327
    # pdfplumber==0.11.6
    # pillow==11.2.1
    # protobuf==6.31.0
    # pyarrow==20.0.0
    # pycparser==2.22
    # pycryptodome==3.23.0
    # pymilvus==2.3.4
    # PyPDF2==3.0.1
    # pypdfium2==4.30.1
    # python-dateutil==2.9.0.post0
    # python-docx==1.1.0
    # python-dotenv==1.1.0
    # pytz==2025.2
    # PyYAML==6.0.2
    # regex==2024.11.6
    # requests==2.32.3
    # safetensors==0.5.3
    # scikit-learn==1.6.1
    # scipy==1.15.3
    # sentence-transformers==2.2.2
    # sentencepiece==0.2.0
    # six==1.17.0
    # sympy==1.14.0
    # threadpoolctl==3.6.0
    # tokenizers==0.21.1
    # torch==2.1.0+cpu
    # torchaudio==2.1.0+cpu
    # torchvision==0.16.0+cpu
    # tqdm==4.67.1
    # transformers==4.52.3
    # typing_extensions==4.13.2
    # tzdata==2025.2
    # ujson==5.10.0
    # urllib3==2.4.0
